update songs in library view
