+++ front-end/app/(tabs)/library/song/[id].tsx
@@ -1,12 +1,13 @@
+import ItemDetailPage from '@/components/shared/ItemDetailPage';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader } from '@/components/ui/card';
 import { useArtistsStore } from '@/stores/artist-store';
 import { useSessionItemsStore } from '@/stores/session-item-store';
 import { useSongsStore } from '@/stores/song-store';
-import { formatTimestampToDate } from '@/utils/date-time';
 import { router, useLocalSearchParams } from 'expo-router';
 import { useEffect, useState } from 'react';
-import { FlatList, Pressable, Text, TextInput, View } from 'react-native';
+import { Pressable, Text, TextInput } from 'react-native';
+import { ScrollView } from 'react-native-gesture-handler';
 
 export default function SongDetailPage() {
   const { id } = useLocalSearchParams<{ id: string }>();
@@ -24,7 +25,7 @@ export default function SongDetailPage() {
 
   const song = songs.find((s) => s.id === id);
   const artist = artists.find((a) => a.id === song?.artist_id);
-  const sessions = sessionItemsBySong[id];
+  const sessionItems = sessionItemsBySong[id] || [];
 
   // Handlers
   const handleBackToArtist = (artistId: string) => {
@@ -39,7 +40,7 @@ export default function SongDetailPage() {
   // Component Layout
   if (!song) return <Text> Song not found! </Text>;
   return (
-    <View className="flex-1 bg-white p-4 space-y-6">
+    <ScrollView className="flex-1 bg-white p-4 space-y-6">
       <Text className="text-2xl font-bold">{song.name}</Text>
 
       {artist && (
@@ -62,36 +63,14 @@ export default function SongDetailPage() {
             keyboardType="numeric"
             className="border border-gray-300 rounded-md p-3 text-base"
           />
-          <Button onPress={handleUpdateGoal}>Update Goal Tempo</Button>
+          <Button onPress={handleUpdateGoal}><Text>Update Goal Tempo</Text></Button>
         </CardContent>
       </Card>
 
       <Text className="text-xl font-semibold">Practice Sessions</Text>
 
-      {sessions && sessions.length === 0 ? (
-        <Text className="text-gray-500 italic">No sessions logged yet.</Text>
-      ) : (
-        <FlatList
-          data={sessions}
-          keyExtractor={(item) => item.id}
-          renderItem={({ item }) => (
-            <Card className="mb-3">
-              <CardHeader className="flex-row justify-between items-center">
-                <Text className="font-medium">{item.tempo} BPM</Text>
-                {/* <View className="flex-row items-center space-x-1">
-                  <Clock size={16} className="text-gray-500" />
-                  <Text className="text-gray-500 text-sm">{item.duration} min</Text>
-                </View> */}
-              </CardHeader>
-              <CardContent>
-                <Text className="text-sm text-gray-400">
-                  {formatTimestampToDate(item.created_at)}
-                </Text>
-              </CardContent>
-            </Card>
-          )}
-        />
-      )}
-    </View>
+      <ItemDetailPage sessionItems={sessionItems} />
+
+    </ScrollView>
   );
 }
